% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dMCMCplot}
\alias{dMCMCplot}
\title{Create a multiple diagnostic plot}
\usage{
dMCMCplot(mcmcobject = NULL, interest = NULL, model = NULL,
  Rprior = NULL, ...)
}
\arguments{
\item{mcmcobject}{a mcmc object (list of samples).}

\item{interest}{a set of parameters of interest. These should have
true priors (e.g. can be directly related to a prior distribution).
Lower-level priors, that depend on a hyperprior, will fail
as dMCMC does not attempt to understand the hierachical
structure of the model file.}

\item{model}{the path to the BUGs/JAGS model text file with
the priors specified. This model must be the one
that generated the MCMC list.}

\item{Rprior}{the prior function directly
specified by the user (as a standard R
probabilty density function e.g. prior=dunif).}

\item{\dots}{additional parameters (unused).
diagnostic plots on. Names must be equal to the names
in the jags/bugs model. If length > 1, one plot is
produced per parameter of interest, in this case output to a
pdf is advised. Otherwise the function forces par(ask=TRUE).}
}
\value{
A multipanel plot
}
\description{
dMCMCplot Create the untimate diagnostic plot for a MCMC chain
}
\details{
Creates a multipanel graphic with diagnostics on parameter samples
}
\examples{
sink("examp.txt")
cat( "
model {
	for (i in 1:N) {
		x[i] ~ dnorm(mu, tau)
	}
	mu ~ dnorm(0, .0001)
	tau <- pow(sigma, -2)
	sigma ~ dunif(0, 100)
    } ",fill=TRUE)
     sink()

jags <- jags.model('examp.txt',
                  data = list('x' = rnorm(100,2,2),
                              'N'=100),
                  n.chains = 4,
                  n.adapt = 100)

mysamples <- coda.samples(jags, c('mu', 'tau'),100)

dMCMCplot(mysamples,"mu","examp.txt")
}
\author{
Marco D. Visser
}
\seealso{
\code{\link{jags.model}}
}
\concept{
MCMC diagnostics
}

