% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findRprior}
\alias{findRprior}
\title{Finds priors and translates these from BUGS to R}
\usage{
findRprior(model = NULL, interest = NULL, silent = FALSE)
}
\arguments{
\item{interest}{a set of parameters to produce}

\item{model}{the path to the BUGs/JAGS model text file with
the priors specified. This model must be the one
that generated the MCMC list.}
}
\description{
findRprior seives through your model files to find and
translate your priors to proper R probabilty density functions.
}
\details{
Given a model file and a the name of the prior
findRprior will return the corresponding
functions in R and translate the parameters from their
BUGS formulation into a R compatable form.
}
\author{
Marco D. Visser
}
\concept{
MCMC diagnostics
}

