% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dMCMCplot}
\alias{dMCMCplot}
\title{Create a multiple diagnostic plot}
\usage{
dMCMCplot(mcmcobject = NULL, interest = NULL, model = NULL,
  Rprior = NULL, ...)
}
\arguments{
\item{mcmcobject}{a mcmc object (list of samples).}

\item{interest}{a set of parameters to produce}

\item{model}{the path to the BUGs/JAGS model text file with
the priors specified. This model must be the one
that generated the MCMC list.}

\item{Rprior}{the prior function directly
specified by the user (as a standard R
probabilty density function e.g. prior=dunif).}

\item{\dots}{additional parameters (unused).
diagnostic plots on. Names must be equal to the names
in the jags/bugs model. If length > 1, one plot is
produced per parameter of interest, in this case output to a
pdf is advised. Otherwise the function forces par(ask=TRUE).}
}
\value{
A multipanel plot
}
\description{
dMCMCplot Create the untimate diagnostic plot for a MCMC chain
}
\details{
Creates a multipanel graphic with diagnostics on parameter samples
}
\examples{
\dontrun{
   # create function
     foo <- function(N){
     N
     }

	}
}
\author{
Marco D. Visser
}
\seealso{
\code{\link{jags.model}}
}
\concept{
MCMC diagnostics
}

